Extract common message component
